fix bazaar item amount

+++ srcs/OpenNos.GameObject/Packets/ClientPackets/CScalcPacket.cs
@@ -16,10 +16,10 @@ namespace OpenNos.GameObject.Packets.ClientPackets
         public short VNum { get; set; }
 
         [PacketIndex(2)]
-        public byte Amount { get; set; }
+        public short Amount { get; set; }
 
         [PacketIndex(3)]
-        public byte MaxAmount { get; set; }
+        public short MaxAmount { get; set; }
 
         [PacketIndex(4)]
         public long Price { get; set; }
